// models/PaginationMeta.ts
/**
 * Telnyx API
 * SIP trunking, SMS, MMS, Call Control and Telephony Data Services.
 *
 * OpenAPI spec version: 2.0.0
 * Contact: support@telnyx.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http.ts';

export class PaginationMeta {
    'totalPages'?: number;
    'totalResults'?: number;
    'pageNumber'?: number;
    'pageSize'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "totalPages",
            "baseName": "total_pages",
            "type": "number",
            "format": ""
        },
        {
            "name": "totalResults",
            "baseName": "total_results",
            "type": "number",
            "format": ""
        },
        {
            "name": "pageNumber",
            "baseName": "page_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "pageSize",
            "baseName": "page_size",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return PaginationMeta.attributeTypeMap;
    }

    public constructor() {
    }
}